pyproject.toml
[tool.poetry]
name = "odh-test"
version = "0.1.0"
description = "testing tool for kserve and odh-kserve (model Controller)"
authors = ["Milind Waykole"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jinja2 = "3.1.4"
kubernetes = "31.0.0"
pytest = "8.0.0"
pyyaml = "6.0.2"
requests = "2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
